fix: separate vulnerabilities from banner service
- Extract vulnerabilities logic from banner service into its own service
- Add Dockerfile for vulnerability service
- Update docker-compose to include vulnerability service
