+++ backend/api_applications/banner_grabbing/main.py
@@ -1,10 +1,6 @@
 import logging
-import os
-import threading
 
-import schedule
-
-from . import db_operations, vulnerability
+from . import db_operations
 
 logging.basicConfig(
     level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
@@ -13,26 +9,7 @@ logging.basicConfig(
 
 def main():
     while True:
-        banners_thread = threading.Thread(
-            target=db_operations.update_banners, name="BannersThread"
-        )
-        vulnerability_thread = threading.Thread(
-            target=db_operations.update_vulnerability, name="VulnerabilityThread"
-        )
-        banners_thread.start()
-        vulnerability_thread.start()
-        logging.info("Waiting for enrichment and banner threads to complete...")
-        banners_thread.join()
-        vulnerability_thread.join()
-        try:
-            schedule.every(1.5).hours.do(
-                vulnerability.download_and_replace_nvd,
-                os.path.join(os.path.dirname(__file__), "cve_data"),
-            )
-        
-        except Exception as e:
-            logging.error("Failed to download and replace CVE file.")
-            
+        db_operations.update_banners()
