+++ backend/api_applications/vulnerability/cve_data/nvdcve-1.1-modified.json.gz
