+++ backend/api_applications/vulnerability/cve_lookup.py
@@ -1,6 +1,5 @@
 import gzip
 import json
-import logging
 import os
 import re
 import shutil
