+++ backend/api_applications/vulnerability/db_operations.py
@@ -1,0 +1,36 @@
+import logging
+
+from shared_libs import monogo_connections
+
+from . import cve_lookup
+
+
+def update_vulnerability():
+
+    try:
+        db = monogo_connections.connect_monogo()
+        results = list(
+            db.scan_results.find(
+                {
+                    "vulnerability": {"$exists": False},
+                    "service_type": {"$exists": True, "$ne": None},
+                    "ports": {"$exists": True, "$nin": [[], None]},
+                }
+            )
+        )
+        try:
+
+            for i in results:
+                vul = cve_lookup.get_vul(i["service_type"])
+                if vul:
+                    logging.info(f"Updating vuls, result is: {vul}")
+                    logging.info(f"getting vuls of : {i}")
+                    db.scan_results.update_one(
+                        {"_id": i["_id"]}, {"$set": {"vulnerability": vul}}
+                    )
+
+        except Exception as e:
+            logging.error(f"Operation do not complete in vuls : {e}")
+
+    except Exception as e:
+        logging.error(f"Operation do not complete in vuls : {e}")
