+++ backend/api_applications/vulnerability/main.py
@@ -1,0 +1,29 @@
+import logging
+import os
+import threading
+
+import schedule
+
+from . import db_operations,cve_lookup
+
+logging.basicConfig(
+    level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
+)
+
+
+def main():
+    while True:
+
+        try:
+            schedule.every(1.5).hours.do(
+                cve_lookup.download_and_replace_nvd,
+                os.path.join(os.path.dirname(__file__), "cve_data"),
+            )
+
+        except Exception as e:
+            logging.error("Failed to download or replace CVE file.")
+        db_operations.update_vulnerability()
+
+
+if __name__ == "__main__":
+    main()
